% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_qc.R
\name{estimate_cor_matrix}
\alias{estimate_cor_matrix}
\title{internal function}
\usage{
estimate_cor_matrix(res.studies, column, cor.method = "spearman")
}
\description{
internal function
}
\keyword{internal}
