% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diff_expr.R
\name{estimate_surrogate_var}
\alias{estimate_surrogate_var}
\title{internal function}
\usage{
estimate_surrogate_var(
  expr,
  pheno,
  var,
  covar = NULL,
  n.sv = 5,
  sv.var.check,
  title
)
}
\description{
internal function
}
\keyword{internal}
