% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_study_results.R
\name{load_study_results}
\alias{load_study_results}
\title{Load study results}
\usage{
load_study_results(info.studies, verbose = TRUE)
}
\arguments{
\item{info.studies}{data.frame with information about each
study (see Details).}

\item{verbose}{Logical. Should additional information be printed? (default:
TRUE).}
}
\value{
list of data.frames with study results
}
\description{
Load results of study specific differential expression analysis.
}
\details{
The data.frame info.studies needs to contain the following columns:
\itemize{
\item id = unique study identifier
\item number.subjects = number of study subjects
\item number.samples = number of study samples (might be identical to
number.subjects)
\item technology = technology (e.g. "RNA-seq" or "array")
\item platform = specific array or sequencing platform
\item file = path to text file with study results
\item column.gene = name of column with gene identifier
\item column.estimate = name of column with estimate (e.g. beta coefficient)
\item column.se = name of column with standard error
\item column.pvalue = name of column with unadjusted P value
\item column.mean.expr = name of column with mean expression value
}
}
