% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diff_expr.R
\name{plot_association_surrogate_var}
\alias{plot_association_surrogate_var}
\title{internal function}
\usage{
plot_association_surrogate_var(pheno, sv.var.check, title)
}
\description{
internal function
}
\keyword{internal}
