% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_qc.R
\name{plot_nr_de_genes}
\alias{plot_nr_de_genes}
\title{QC on number of differentially expressed genes depending on sample size}
\usage{
plot_nr_de_genes(info.studies, res.studies)
}
\arguments{
\item{info.studies}{[data.frame] Information about studies (need to include
a column called "id" with a unique study identifier and a column 
"number.subjects" specifying the number of independent subjects in the study.}

\item{res.studies}{[list] Study specific
results (as generated by \code{\link{load_study_results}}.}
}
\value{
scatterplot as object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Differentially expressed (DE) genes in each study are determined based on
adjusted P values < 0.05. For comparability only the number of 
DE genes among the genes available in each of the studies will be plotted
depending on the number of subjects.
}
