% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_diff_expr.R
\name{plot_variance_explained_surrogate_var}
\alias{plot_variance_explained_surrogate_var}
\title{internal function}
\usage{
plot_variance_explained_surrogate_var(svaobj, expr, title = "")
}
\description{
based on code from
https://support.bioconductor.org/p/88553/
}
\keyword{internal}
