% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_qc.R
\name{qc_diff_expr_genes}
\alias{qc_diff_expr_genes}
\title{QC on genes differentially expressed in each study}
\usage{
qc_diff_expr_genes(res.studies, group = NULL)
}
\arguments{
\item{res.studies}{[list] Study specific
results (as generated by \code{\link{load_study_results}}.}

\item{group}{[character(n) or numeric(n)] groups which should be used for
stratification in boxplot (e.g. information about technology per study)}
}
\value{
data.frame with gene level statistics for each study:
\itemize{
\item id = unique study identifier
\item number.diff.expr.genes = number of DE genes
\item number.diff.expr.genes.outlier = outlier based on
number.diff.expr.genes (1: yes, 0: no, NA: not calculated due to small
number of studies)
\item freq.up.genes = relative frequency of DE genes that are upregulated
\item freq.up.genes.outlier = outlier based on
freq.up.genes (1: yes, 0: no, NA: not calculated due to small
number of studies)
\item median.overlapping.diff.expr.genes = median number of overlapping DE
and consistent genes with other studies
\item median.overlapping.diff.expr.genes.outlier = outlier based on
median.overlapping.diff.expr.genes (1: yes, 0: no, NA: not calculated due to small
number of studies)
}
}
\description{
Differentially expressed genes in each study are determined based on
adjusted P values < 0.05. Calculate total number of differentially expressed
(DE) genes per study, frequency of DE genes that are upregulated and number
of studies covering each DE gene (not taking into account direction of
effect) and numbers of overlapping DE and consistent genes in each pair of
studies.
}
