% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_qc.R
\name{qc_genes}
\alias{qc_genes}
\title{QC on genes available in each study}
\usage{
qc_genes(res.studies, group = NULL)
}
\arguments{
\item{res.studies}{[list] study specific
results (as generated by \code{\link{load_study_results}}.}

\item{group}{[character(n) or numeric(n)] groups which should be used for
stratification in boxplot (e.g. information about technology per study)}
}
\value{
data.frame with gene level statistics for each study:
\itemize{
\item id = unique study identifier
\item number.genes = total number of genes
\item median.overlapping.genes = median number of overlapping genes with
other studies
\item median.overlapping.genes.outlier = outlier based on
median.overlapping.genes (1: yes, 0: no, NA: not calculated due to small
number of studies)
}
}
\description{
Calculate total number of genes per study, number of studies covering each
gene and numbers of overlapping genes in each pair of studies.
}
