% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_qc.R
\name{qc_pvalues}
\alias{qc_pvalues}
\title{QC on P-values}
\usage{
qc_pvalues(res.studies)
}
\arguments{
\item{res.studies}{[list] study specific
results (as generated by \code{\link{load_study_results}}.}
}
\description{
Histogram of all P-values for each study.
}
